chore: combine the update metrics scripts.
