+++ scripts/update_metrics.ts
@@ -7,16 +7,23 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 
-import type {ParsedArguments} from '../build/src/bin/chrome-devtools-mcp-cli-options.js';
 import {
+  cliOptions,
+  parseArguments,
+} from '../build/src/bin/chrome-devtools-mcp-cli-options.js';
+import {
+  getPossibleFlagMetrics,
+  type FlagMetric,
+} from '../build/src/telemetry/flagUtils.js';
+import {
+  applyToExisting,
   applyToExistingMetrics,
   generateToolMetrics,
   type ToolMetric,
 } from '../build/src/telemetry/toolMetricsUtils.js';
-import type {ToolDefinition} from '../build/src/tools/ToolDefinition.js';
 import {createTools} from '../build/src/tools/tools.js';
 
-export function HaveUniqueNames(tools: ToolDefinition[]): boolean {
+export function HaveUniqueNames(tools: Array<{name: string}>): boolean {
   const toolNames = tools.map(tool => tool.name);
   const toolNamesSet = new Set(toolNames);
   return toolNamesSet.size === toolNames.length;
@@ -30,8 +37,9 @@ function writeToolCallMetricsConfig() {
     throw new Error(`Error: Directory ${dir} does not exist.`);
   }
 
-  const fullTools = createTools({slim: false} as ParsedArguments);
-  const slimTools = createTools({slim: true} as ParsedArguments);
+  // Avoid 'as ParsedArguments' by using parseArguments
+  const fullTools = createTools(parseArguments('0.0.0', ['', '']));
+  const slimTools = createTools(parseArguments('0.0.0', ['', '', '--slim']));
 
   const allTools = [...fullTools, ...slimTools];
 
@@ -62,4 +70,43 @@ function writeToolCallMetricsConfig() {
   );
 }
 
-writeToolCallMetricsConfig();
+function writeFlagUsageMetrics() {
+  const outputPath = path.resolve('src/telemetry/flag_usage_metrics.json');
+
+  const dir = path.dirname(outputPath);
+  if (!fs.existsSync(dir)) {
+    throw new Error(`Error: Directory ${dir} does not exist.`);
+  }
+
+  let existingMetrics: FlagMetric[] = [];
+  if (fs.existsSync(outputPath)) {
+    try {
+      existingMetrics = JSON.parse(
+        fs.readFileSync(outputPath, 'utf8'),
+      ) as FlagMetric[];
+    } catch {
+      console.warn(
+        `Warning: Failed to parse existing metrics from ${outputPath}. Starting fresh.`,
+      );
+    }
+  }
+
+  const newMetrics = getPossibleFlagMetrics(cliOptions);
+  const mergedMetrics = applyToExisting<FlagMetric>(
+    existingMetrics,
+    newMetrics,
+  );
+
+  fs.writeFileSync(outputPath, JSON.stringify(mergedMetrics, null, 2) + '\n');
+
+  console.log(
+    `Successfully wrote ${mergedMetrics.length} flag usage metrics to ${outputPath}`,
+  );
+}
+
+function main() {
+  writeToolCallMetricsConfig();
+  writeFlagUsageMetrics();
+}
+
+main();
